+++ controllers/modules/busboy.js
@@ -86,11 +86,11 @@ function ImgUploadByBusboy (req, res, limits, next) {
     // BUSBOY-LISTENER: after all streams resolved
     busboy.on('finish', () => {
         req.body.busboySlip = { raw: Object.values(populator).filter(obj => obj.isSkipped !== true), notice };
-        next();
+        return next();
     });
 
     // ACTIVATION: pipe busboy
-    req.pipe(busboy);
+    return req.pipe(busboy);
 }