+++ controllers/render.js
@@ -20,14 +20,16 @@ exports.post = (view, doc) => async (req, res) => {
 
 
 
-// error handler branches
+// error handler
+// branches
 const branch = {};
 branch.MongoError = (err, req, res, next) => {
     if (err.code === 11000) req.flash('error', 'This username is not available.');
     return res.redirect('back');
 };
 
-// error handler core
+
+// core
 exports.errorHandler = (err, req, res, next) => {     // todo: error handler separations
     const { MongoError } = require('mongodb');
     if (err instanceof MongoError) return branch.MongoError(err, req, res, next);
@@ -38,5 +40,6 @@ exports.errorHandler = (err, req, res, next) => {     // todo: error handler sep
     }
 
     req.flash('error', err.toString());
-    res.redirect('/');
+    if (process.env.NODE_ENV === 'dev') console.log(err);
+    return res.redirect('/');
 };