+++ controllers/router/dashboard.js
@@ -21,7 +21,7 @@ dashboard.setting = {
     get: (req, res) => res.render('./dashboard/setting'),
     patch: async (req, res) => {
         await settingModel.updateSettings(req.body.siteSetting);
-        res.redirect('back');
+        return res.redirect('back');
     },
 };
 
@@ -29,7 +29,7 @@ dashboard.profile = {
     get: (req, res) => res.render('./dashboard/account/profile'),
     patch: async (req, res) => {
         await userModel.update({ _id: req.user._id }, { $set: req.body.profile });
-        res.redirect('back');
+        return res.redirect('back');
     },
 };
 
@@ -38,7 +38,7 @@ dashboard.security = {
     patch: [_md.passwordValidation, async (req, res) => {
         await req.user.changePassword(req.body.password.old, req.body.password.new);
         req.flash('info', 'Password have been successfully changed.');
-        res.redirect('back');
+        return res.redirect('back');
     }],
 };
 
@@ -51,6 +51,6 @@ dashboard.upload = {   // todo: to be integrated in profile and media manager
         }
         const doc = await mediaModel.mediaCreateThenAssociate(req.body.busboySlip.raw, req.user);
         if (doc.length > 0) req.flash('info', `${doc.length} file(s) successfully uploaded!`);
-        res.redirect('back');
+        return res.redirect('back');
     }],
 };