+++ routers/index.js
@@ -23,16 +23,16 @@ function _useMiddleware(app) {
 
 function _useRoutes(app) {
     // seed
-    if (process.env.NODE_ENV === 'dev' || 'test') app.use('/seed', require('../routes/seed'));
+    if (process.env.NODE_ENV === 'dev' || 'test') app.use('/seed', require('.//seed'));
 
     // units
-    app.use('/dashboard', require('../routes/dashboard'));
-    app.use('/post', require('../routes/post'));
-    app.use('/', require('../routes/authentication'));
-    app.use('/', require('../routes/page'));
+    app.use('/dashboard', require('.//dashboard'));
+    app.use('/post', require('.//post'));
+    app.use('/', require('.//authentication'));
+    app.use('/', require('.//page'));
 
     // error
-    app.use('/', require('../routes/error'));
+    app.use('/', require('.//error'));
 }